+++ xmlschema/__init__.py
@@ -8,6 +8,7 @@
 #
 # @author Davide Brunato <brunato@sissa.it>
 #
+import warnings
 from .exceptions import XMLSchemaException, XMLSchemaRegexError, XMLSchemaURLError
 from .resources import (
     normalize_url, fetch_resource, load_xml_resource, fetch_namespaces,
@@ -25,13 +26,17 @@ from .validators import (
     XMLSchemaImportWarning, XsdGlobals, XMLSchemaBase, XMLSchema, XMLSchema10
 )
 
-__version__ = '1.0.8'
+__version__ = '1.0.9'
 __author__ = "Davide Brunato"
 __contact__ = "brunato@sissa.it"
 __copyright__ = "Copyright 2016-2019, SISSA"
 __license__ = "MIT"
 __status__ = "Production/Stable"
 
-# Backward compatibility (will be removed in 1.1 version)
-XMLSchema_v1_0 = XMLSchema10
+
+# API deprecation warnings
+def XMLSchema_v1_0():
+    warnings.warn("XMLSchema_v1_0 will be removed in 1.1 version, use XMLSchema10 instead", DeprecationWarning)
+    return XMLSchema10
+
 etree_get_namespaces = fetch_namespaces