+++ xmlschema/tests/test_all.py
@@ -9,35 +9,18 @@
 #
 # @author Davide Brunato <brunato@sissa.it>
 #
-"""
-This module runs all tests of the 'xmlschema' package.
-"""
+
 if __name__ == '__main__':
-    import unittest
-    import os
-    import sys
-
-    try:
-        import xmlschema
-    except ImportError:
-        # Adds the package base dir path as first search path for imports
-        pkg_base_dir = os.path.dirname(os.path.dirname(os.path.dirname(__file__)))
-        sys.path.insert(0, pkg_base_dir)
-        import xmlschema
-
-    from xmlschema.tests import print_test_header, tests_factory
-    from xmlschema.tests.test_helpers import TestHelpers
-    from xmlschema.tests.test_meta import TestXsd10BuiltinTypes, TestXsd11BuiltinTypes, TestGlobalMaps
-    from xmlschema.tests.test_regex import TestCodePoints, TestUnicodeSubset, TestUnicodeCategories, TestPatterns
-    from xmlschema.tests.test_xpath import XsdXPathTest
-    from xmlschema.tests.test_resources import TestResources
-    from xmlschema.tests.test_models import TestModelValidation
-    from xmlschema.tests.test_schemas import make_schema_test_class, TestXMLSchema10, TestXMLSchema11
-    from xmlschema.tests.test_validators import make_validator_test_class, TestValidation, TestDecoding, \
-        TestDecoding11, TestEncoding, TestEncoding11
-    from xmlschema.tests.test_package import TestPackaging
+    from xmlschema.tests.test_helpers import *
+    from xmlschema.tests.test_meta import *
+    from xmlschema.tests.test_regex import *
+    from xmlschema.tests.test_xpath import *
+    from xmlschema.tests.test_resources import *
+    from xmlschema.tests.test_models import *
+    from xmlschema.tests.test_schemas import *
+    from xmlschema.tests.test_validators import *
+    from xmlschema.tests.test_package import *
+    from xmlschema.tests import print_test_header
 
     print_test_header()
-    globals().update(tests_factory(make_schema_test_class, 'xsd'))
-    globals().update(tests_factory(make_validator_test_class, 'xml'))
     unittest.main()